netplan:parse-nm: rename generator/parser modules
